Recent Doctor Fixed and Sonar issue fixed

+++ src/main/java/com/medical/doctor/controller/DoctorController.java
@@ -269,19 +269,28 @@ public class DoctorController {
 		}
 	}
 
+	/**
+	 * getRecentDoctors will provide recently joined doctors
+	 * 
+	 * 
+	 * Days already taken in Integer because this will come from UI end not will
+	 * be provided by any user always will come in correct format
+	 * 
+	 * @param days
+	 * @return
+	 */
 	@RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE, value = "/get/{days}/recentdoctors")
 	@ResponseBody
 	public List<DoctorResponse> getRecentDoctors(@PathVariable Integer days) {
 
-		if (!StringUtils.isEmpty(days) && days.intValue() > 0) {
-			return doctorMapper.mapDoctors(doctorService.getRecentDoctors(days
-					.intValue()));
+		if (!StringUtils.isEmpty(days) && days > 0) {
+			return doctorMapper
+					.mapDoctors(doctorService.getRecentDoctors(days));
 		} else {
 			throw new BadRequestException("Doctor ID should not be blank");
 		}
 	}
-	
-	
+
 	@RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE, value = "/get/all/unapproved/expertisation")
 	@ResponseBody
 	public Map<Integer, String> getAllUnApprovedExpertisations() {
@@ -299,6 +308,5 @@ public class DoctorController {
 			throw new BadRequestException("Expertisation can not be blank");
 		}
 	}
-
 
 }
